[Feat] 주문-배송-배송담당자 지정 서비스 연동 및 테스트 (#101)
* [refact] 주문 디렉토리 구조 DDD 설계에 맞게 재구성

* [refact] 주문 서비스 에러처리 고도화

* [refact] 배송 디렉토리 구조 DDD 설계에 맞게 재구성

* [refact] redis 의존성 제거

* [refact] 배송 서비스 에러처리 고도화

* [feat] 배송 서비스 서킷브레이커 의존성 추가

* [feat] 배송-배송담당자 할당 서비스 연동

* [feat] 배송 서비스 fallbackFactory 클래스 도입

+++ msa.order/src/main/java/com/msa/order/presentation/response/ReadOrderResponse.java
@@ -17,7 +17,7 @@ public record ReadOrderResponse(
     OrderStatus status,
     UUID itemId,
     String itemName,
-    int quantity,
+    Long quantity,
     String description,
     LocalDateTime createdAt,
     LocalDateTime updatedAt,
